+++ src/main/kotlin/no/nav/meldeplikt/meldekortservice/api/Person.kt
@@ -34,6 +34,7 @@ fun Routing.personApi(httpClient: HttpClient) {
     getMeldekort()
     kontrollerMeldekort()
     endreMeldeform()
+    pingWeblogic()
 }
 
 private const val personGroup = "Person"
@@ -124,4 +125,19 @@ fun Routing.endreMeldeform() =
         )
     ) {_, meldeform ->
         call.respond(status=HttpStatusCode.OK,message = "Meldeform er ikke implementert")
+    }
+
+@Group(personGroup)
+@Location("$PERSON_PATH/weblogic")
+class PingWeblogic
+
+//Pinger weblogic
+fun Routing.pingWeblogic() =
+    get<PingWeblogic>(
+        "Ping weblogic".securityAndReponds(
+            BearerTokenSecurity(),
+            ok<String>(),
+            serviceUnavailable<ErrorMessage>(),
+            unAuthorized<Error>())) {
+        call.respond(status = HttpStatusCode.OK, message = "${SoapConfig.oppfoelgingPing()}")
     }
